Issues with preloading resources
Referencing resources contained in NewActorReferences fails if NewActorReferences is referenced as a preload or if it is instantiated and then referenced, but it succeeds if NewActorReferences is a global singleton.

I don't really understand why that is, but whatever. I can make all of the resource dictionaries global singletons, but then I'd have way too many globals. I think I'll put all resources in one huge file and then make global singletons to access them in an organised and sensible manner.

Just a bunch of overhead that I have to do for no reason...